storyteller/requirements.txt
Flask
pyttsx3
python-dotenv
langchain
groq
langchain_groq
gunicorn